refactor: moved repeated code to functions
