+++ utils/misc.mjs
@@ -1,0 +1,2 @@
+export const sleep = async (dur) =>
+  await new Promise((resolve) => setTimeout(resolve, dur));
